+++ app/src/main/java/com/example/StoreAnalytics.java
@@ -18,7 +18,7 @@ public class StoreAnalytics {
      * @return A list of entries containing the product and the number of times it
      *         was ordered, sorted by the number of orders in descending order.
      */
-    public List<Entry<Product, Double>> getMostPopularProducts() {
+    public List<Entry<Product, Integer>> getMostPopularProducts() {
         // We start with remapping all of out orders to a map of Product and how many of
         // them has been orderer.
         return orderManager.getOrders()
@@ -35,16 +35,16 @@ public class StoreAnalytics {
                 .flatMap(p -> p.entrySet().stream()) // Stream<Entry<Product, Integer>>
 
                 // Merge together all entries that share key, adding their values together.
-                .collect(Collectors.toMap(Entry::getKey, e -> e.getValue().doubleValue(), Double::sum))
+                .collect(Collectors.toMap(Entry::getKey, e -> e.getValue().intValue(), Integer::sum))
 
-                // At this point we have a Map<Product, Double> where the double is the number
+                // At this point we have a Map<Product, Integer> where the integer is the number
                 // of times
                 // that product has been ordered.
                 .entrySet()
-                .stream() // Stream<Entry<Product, Double>>
+                .stream() // Stream<Entry<Product, Integer>>
 
                 // Sort the stream based on the Entry values, since this represents the amount
-                .sorted(Entry.<Product, Double>comparingByValue(Comparator.reverseOrder()))
+                .sorted(Entry.<Product, Integer>comparingByValue(Comparator.reverseOrder()))
                 .toList();
     }
 
@@ -55,7 +55,7 @@ public class StoreAnalytics {
      * @return A list of entries containing the product and the number of times it
      *         was ordered, sorted by the number of orders in descending order.
      */
-    public List<Entry<Product, Double>> getMostPopularProducts(int topN) {
+    public List<Entry<Product, Integer>> getMostPopularProducts(int topN) {
         // Implemented using the method that returns all products.
         return getMostPopularProducts()
                 // Return only the top N products, or all if there are less than N.
